smart_db_csv_builder/requirements.txt
# Dataset Builder — Backend Dependencies
# Install with: pip install -r requirements.txt

# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Database drivers (install only what you need)
pymysql==1.1.1           # MySQL / MariaDB
psycopg2-binary==2.9.9   # PostgreSQL
pymongo==4.7.2           # MongoDB
pyodbc==5.1.0            # MSSQL (requires ODBC driver installed on OS)
# sqlite3 is built into Python stdlib

# Data processing
pandas==2.2.2
numpy==1.26.4

# LLM
anthropic==0.28.0        # Claude API (optional — falls back to Ollama if not provided)

# Utilities
python-dotenv==1.0.1
pydantic==2.7.1
